Treci dan - prvi komit

+++ PROJEKTNIKOD/PROJEKTNIKOD/Student.cpp
@@ -102,6 +102,9 @@ void Student::sendFriendRequest(std::string userName) const {
 			throw std::exception("Person to whom you want to send friend request doesn't exist!");
 		}
 		getFriendFile.close();
+		if (checkIfFriendRequestIsAlreadyMade(userName)) {
+			throw std::exception("You've already made a request, wait!");
+		}
 		auto openFriendFile = std::ofstream("./STUDENTI/" + userName + "/" + "FRIENDREQUESTS.TXT", std::ios::out | std::ios::app);
 		if (openFriendFile) {
 			openFriendFile << *this;
@@ -113,11 +116,28 @@ void Student::sendFriendRequest(std::string userName) const {
 	}
 }
 
-void Student::automaticLecturerFriend(std::string s1, std::string s2) const {
+bool Student::checkIfFriendRequestIsAlreadyMade(std::string userName) const {
+	auto getFriendFile = std::ifstream("./STUDENTI/" + userName + "/" + "FRIENDREQUESTS.TXT", std::ios::in);
+	if (getFriendFile) {
+		while (getFriendFile.good()) {
+			User u;
+			getFriendFile >> u;
+			if (u.getUserName() == userName) {
+				getFriendFile.close();
+				return true;
+			}
+		}
+	}
+
+	getFriendFile.close();
+	return false;
+}
+
+void Student::automaticLecturerFriend(std::string lecturerName, std::string lecturerPassword) const {
 	auto writeFriend = std::ofstream("./STUDENTI/" + this->userName + "/FRIENDS.txt", std::ios::out | std::ios::app);
-	auto writeLecturerFriend = std::ofstream("./STUDENTI/" + s1 + "/FRIENDS.txt", std::ios::out | std::ios::app);
+	auto writeLecturerFriend = std::ofstream("./STUDENTI/" + lecturerName + "/FRIENDS.txt", std::ios::out | std::ios::app);
 	if (writeFriend) {
-		Lecturer l(s1, s2);
+		Lecturer l(lecturerName, lecturerPassword);
 		writeFriend << l;
 		writeLecturerFriend << *this;
 	}
@@ -135,33 +155,36 @@ void Student::checkFriendRequests() const {
 			while (openFriendRequests.good()) {
 				User u;
 				openFriendRequests >> u;
-				users.push_back(u);
-				std::cout << u;
+				if (std::find(users.begin(), users.end(), u) == users.end()) {
+					users.push_back(u);
+					std::cout << u;
+				}
 			}
 		}
 		openFriendRequests.close();
 
 		std::cout << "Do you want any of them to become your friends: (y/n)" << std::endl;
 		std::string choice;
-		std::cin >> choice;
+		std::getline(std::cin, choice, '\n');
 
 		if (choice == "y") {
 			std::cout << "What's the name of the student you wanna get friends with: " << std::endl;
 			std::string name;
+			//std::cin >> name;
 			std::getline(std::cin, name, '\n');
 
 			if (this->checkIfIsAlreadyAFriend(name)) {
 				throw std::exception("This person is already your friend!");
 			}
 
-			auto friendFile = std::ofstream("./STUDENTI/" + this->userName + "/" + "FRIENDS.TXT", std::ios::out | std::ios::app);
+			auto friendFile = std::ofstream("./STUDENTI/" + this->userName + "/FRIENDS.TXT", std::ios::out | std::ios::app);
 
 			//auto friendRequestFile = std::ifstream("./STUDENTI/" + this->userName + "/" + "FRIENDREQUESTS.TXT", std::ios::in);
 			if (friendFile) {
 				for (auto elem : users) {
 					if (elem.getUserName() == name && elem.getUserName() != "") {
 						friendFile << elem;
-						auto friendFile2 = std::ofstream("./STUDENTI/" + name + "/" + "FRIENDS.TXT", std::ios::out | std::ios::app);
+						auto friendFile2 = std::ofstream("./STUDENTI/" + name + "/FRIENDS.TXT", std::ios::out | std::ios::app);
 						friendFile2 << *this;
 					}
